Create countingSort
Counting sort takes in an array of integers in random order and counts the number of times that value shows up in the array. Using the length of the data pushed in to the array adds the length value to a new array and then returns the array. Uses linear search.